--- src/components/PostPage.tsx
@@ -1,15 +1,0 @@
-import PostDetail from "pages/posts/detail";
-import Header from "components/Header";
-import Footer from "components/Footer";
-
-const PostPage = () => {
-  return (
-    <>
-      <Header />
-      <PostDetail />
-      <Footer />
-    </>
-  );
-};
-
-export default PostPage;
