+++ src/index.tsx
@@ -5,7 +5,7 @@ import "index.css";
 
 import Home from "pages/home";
 import PostList from "pages/posts";
-import PostPage from "components/PostPage";
+import PostPage from "pages/posts/detail";
 import PostNew from "pages/posts/new";
 import PostEdit from "pages/posts/edit";
 import ProfilePage from "components/ProfilePage";
